Debugging

+++ src/main/java/ch/ethz/matsim/av/framework/AVConfigGroup.java
@@ -29,7 +29,7 @@ public class AVConfigGroup extends ReflectiveConfigGroup {
 	}
 
 	@StringGetter(PARALLEL_ROUTERS)
-	public double getParallelRouters() {
+	public long getParallelRouters() {
 		return parallelRouters;
 	}
